requirements.txt
aiosmtplib==3.0.2
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cssselect==1.3.0
dnspython==2.7.0
email_validator==2.2.0
fake-useragent==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
fastapi-mail==1.5.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.4.0
lxml_html_clean==0.4.2
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.10.18
parse==1.20.2
passlib==1.7.4
playwright==1.52.0
pydantic==2.11.5
pydantic-extra-types==2.10.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.1
PyJWT==2.10.1
PyMySQL==1.1.1
pyquery==2.0.1
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.7
shellingham==1.5.4
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.46.2
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
ujson==5.10.0
urllib3==1.26.20
uvicorn==0.34.3
w3lib==2.3.1
watchfiles==1.0.5
websockets==10.4
zipp==3.23.0
